Change pow function to use long long type
Before,this function is return an int type number,but the"res"and"po" are long long type.

+++ docs/math/code/inverse/inverse-2.cpp
@@ -2,7 +2,7 @@
 
 // --8<-- [start:core]
 // Binary exponentiation.
-int pow(int a, int b, int m) {
+long long pow(long long a, long long b, long long m) {
   long long res = 1, po = a;
   for (; b; b >>= 1) {
     if (b & 1) res = res * po % m;
